requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pandas
qrcode
fpdf
reportlab
opencv-python-headless
pillow==9.5.0
aiofiles
openpyxl